requirements.txt
pyyaml>=6.0
colorlog>=6.7.0
click>=8.1.0
python-dateutil>=2.8.2
typing-extensions>=4.0.0 